// Frontend-production-status/src/components/Theme/type.ts
export type Theme =
    | "light"
    | "dark"
    | "cupcake"
    | "bumblebee"
    | "emerald"
    | "corporate"
    | "synthwave"
    | "retro"
    | "cyberpunk"
    | "valentine"
    | "halloween"
    | "garden"
    | "forest"
    | "aqua"
    | "lofi"
    | "pastel"
    | "fantasy"
    | "wireframe"
    | "black"
    | "luxury"
    | "dracula"
    | "cmyk"
    | "autumn"
    | "business"
    | "acid"
    | "lemonade"
    | "night"
    | "coffee"
    | "winter"
    | "dim"
    | "nord"
    | "sunset"
    | "caramellatte"
    | "abyss"
    | "silk"
    | "custom"
    | "calm-day"
    | 'Dark-Purple'